Fix warning: implicit conversion changes signedness
visp/tutorial/tracking/render-based/tutorial-rbt-sequence.cpp: In function ‘int main(int, const char**)’:
visp/tutorial/tracking/render-based/tutorial-rbt-sequence.cpp:136:39: warning: conversion to ‘int’ from ‘unsigned int’ may change the sign of the result [-Wsign-conversion]
  136 |   const int width = readerRGB.getWidth();
      |                     ~~~~~~~~~~~~~~~~~~^~
visp/tutorial/tracking/render-based/tutorial-rbt-sequence.cpp:137:41: warning: conversion to ‘int’ from ‘unsigned int’ may change the sign of the result [-Wsign-conversion]
  137 |   const int height = readerRGB.getHeight();
      |                      ~~~~~~~~~~~~~~~~~~~^~
visp/tutorial/tracking/render-based/tutorial-rbt-sequence.cpp:141:29: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
  141 |   vpImage<unsigned char> Id(height, width);
      |                             ^~~~~~
visp/tutorial/tracking/render-based/tutorial-rbt-sequence.cpp:141:37: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
  141 |   vpImage<unsigned char> Id(height, width);
      |                                     ^~~~~
visp/tutorial/tracking/render-based/tutorial-rbt-sequence.cpp:143:39: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
  143 |   vpImage<unsigned char> depthDisplay(height, width);
      |                                       ^~~~~~
visp/tutorial/tracking/render-based/tutorial-rbt-sequence.cpp:143:47: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
  143 |   vpImage<unsigned char> depthDisplay(height, width);
      |                                               ^~~~~
visp/tutorial/tracking/render-based/tutorial-rbt-sequence.cpp:144:40: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
  144 |   vpImage<unsigned char> IProbaDisplay(height, width);
      |                                        ^~~~~~
visp/tutorial/tracking/render-based/tutorial-rbt-sequence.cpp:144:48: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
  144 |   vpImage<unsigned char> IProbaDisplay(height, width);
      |                                                ^~~~~
visp/tutorial/tracking/render-based/tutorial-rbt-sequence.cpp:145:27: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
  145 |   vpImage<vpRGBa> IRender(height, width);
      |                           ^~~~~~
visp/tutorial/tracking/render-based/tutorial-rbt-sequence.cpp:145:35: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
  145 |   vpImage<vpRGBa> IRender(height, width);
      |                                   ^~~~~
visp/tutorial/tracking/render-based/tutorial-rbt-sequence.cpp:146:32: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
  146 |   vpImage<vpRGBa> InormDisplay(height, width);
      |                                ^~~~~~
visp/tutorial/tracking/render-based/tutorial-rbt-sequence.cpp:146:40: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
  146 |   vpImage<vpRGBa> InormDisplay(height, width);
      |                                        ^~~~~
visp/tutorial/tracking/render-based/tutorial-rbt-sequence.cpp:147:40: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
  147 |   vpImage<unsigned char> ICannyDisplay(height, width);
      |                                        ^~~~~~
visp/tutorial/tracking/render-based/tutorial-rbt-sequence.cpp:147:48: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
  147 |   vpImage<unsigned char> ICannyDisplay(height, width);
      |                                                ^~~~~
visp/tutorial/tracking/render-based/tutorial-rbt-sequence.cpp:203:97: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
  203 |         std::string depthName = vpIoTools::formatString(sequenceArgs.depthFolder + "/%06d.npy", im);
      |                                                                                                 ^~

+++ tutorial/tracking/render-based/tutorial-rbt-sequence.cpp
@@ -1,6 +1,6 @@
 /*
  * ViSP, open source Visual Servoing Platform software.
- * Copyright (C) 2005 - 2024 by Inria. All rights reserved.
+ * Copyright (C) 2005 - 2026 by Inria. All rights reserved.
  *
  * This software is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -54,7 +54,7 @@ using namespace VISP_NAMESPACE_NAME;
 struct CmdArguments
 {
   CmdArguments() : startFrame(-1), frameStep(1), stepByStep(false)
-  { }
+  {}
 
   void registerArguments(vpJsonArgumentParser &parser)
   {
@@ -133,8 +133,8 @@ int main(int argc, const char **argv)
     std::cout << "  First frame: " << readerRGB.getFirstFrameIndex() << std::endl;
   }
 
-  const int width = readerRGB.getWidth();
-  const int height = readerRGB.getHeight();
+  const unsigned int width = readerRGB.getWidth();
+  const unsigned int height = readerRGB.getHeight();
 
   std::cout << "  Image size : " << width << " x " << height << std::endl;
 
@@ -200,7 +200,7 @@ int main(int argc, const char **argv)
       readerRGB.acquire(Icol);
       vpImageConvert::convert(Icol, Id);
       if (!sequenceArgs.depthFolder.empty()) {
-        std::string depthName = vpIoTools::formatString(sequenceArgs.depthFolder + "/%06d.npy", im);
+        std::string depthName = vpIoTools::formatString(sequenceArgs.depthFolder + "/%06d.npy", static_cast<unsigned int>(im));
         visp::cnpy::NpyArray npz_data = visp::cnpy::npy_load(depthName);
         vpImage<uint16_t> dataArray(npz_data.data<uint16_t>(), npz_data.shape[0], npz_data.shape[1], false);
         float scale = 9.999999747378752e-05;
